update to latest idea for args
