--- src/main/java/gov/nasa/pds/harvest/cfg/CognitoType.java
@@ -1,149 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v4.0.3 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-//
-
-
-package gov.nasa.pds.harvest.cfg;
-
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlAttribute;
-import jakarta.xml.bind.annotation.XmlSchemaType;
-import jakarta.xml.bind.annotation.XmlType;
-import jakarta.xml.bind.annotation.XmlValue;
-import jakarta.xml.bind.annotation.adapters.NormalizedStringAdapter;
-import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
-
-
-/**
- * 
- *         Currently, cognito to serverless opensearh requires three bits of
- *         information:
- *           1. clientID: magic ID which must be given and is not a secret but
- *                        sure looks like one
- *           2. @gateway: an endpoint that exchanges the cognito authentication
- *                        to a signed URL that can be used for communicating with
- *                        opensearch on AWS
- *           3. @IDP: the cognito authentication endpoint that exchanges the
- *                    username/password to a set of tokens used by the gateway.
- *       
- * 
- * <p>Java class for cognito_type complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>{@code
- * <complexType name="cognito_type">
- *   <simpleContent>
- *     <extension base="<http://www.w3.org/2001/XMLSchema>normalizedString">
- *       <attribute name="gateway" type="{http://www.w3.org/2001/XMLSchema}normalizedString" default="https://c8u1zk30u5.execute-api.us-west-2.amazonaws.com/dev/signed-url" />
- *       <attribute name="IDP" type="{http://www.w3.org/2001/XMLSchema}normalizedString" default="cognito-idp.us-west-2.amazonaws.com" />
- *     </extension>
- *   </simpleContent>
- * </complexType>
- * }</pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "cognito_type", propOrder = {
-    "value"
-})
-public class CognitoType {
-
-    @XmlValue
-    @XmlJavaTypeAdapter(NormalizedStringAdapter.class)
-    @XmlSchemaType(name = "normalizedString")
-    protected String value;
-    @XmlAttribute(name = "gateway")
-    @XmlJavaTypeAdapter(NormalizedStringAdapter.class)
-    @XmlSchemaType(name = "normalizedString")
-    protected String gateway;
-    @XmlAttribute(name = "IDP")
-    @XmlJavaTypeAdapter(NormalizedStringAdapter.class)
-    @XmlSchemaType(name = "normalizedString")
-    protected String idp;
-
-    /**
-     * Gets the value of the value property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getValue() {
-        return value;
-    }
-
-    /**
-     * Sets the value of the value property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setValue(String value) {
-        this.value = value;
-    }
-
-    /**
-     * Gets the value of the gateway property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getGateway() {
-        if (gateway == null) {
-            return "https://c8u1zk30u5.execute-api.us-west-2.amazonaws.com/dev/signed-url";
-        } else {
-            return gateway;
-        }
-    }
-
-    /**
-     * Sets the value of the gateway property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setGateway(String value) {
-        this.gateway = value;
-    }
-
-    /**
-     * Gets the value of the idp property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getIDP() {
-        if (idp == null) {
-            return "cognito-idp.us-west-2.amazonaws.com";
-        } else {
-            return idp;
-        }
-    }
-
-    /**
-     * Sets the value of the idp property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setIDP(String value) {
-        this.idp = value;
-    }
-
-}
