+++ src/main/java/gov/nasa/pds/harvest/cfg/RegistryType.java
@@ -10,9 +10,9 @@ package gov.nasa.pds.harvest.cfg;
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
 import jakarta.xml.bind.annotation.XmlAttribute;
-import jakarta.xml.bind.annotation.XmlElement;
 import jakarta.xml.bind.annotation.XmlSchemaType;
 import jakarta.xml.bind.annotation.XmlType;
+import jakarta.xml.bind.annotation.XmlValue;
 import jakarta.xml.bind.annotation.adapters.NormalizedStringAdapter;
 import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 
@@ -20,15 +20,12 @@ import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 /**
  * 
  *         Define the connection to the registry, security for the connection, and
- *         the index within the registry.
+ *         the index within the registry. The value of this tag is a pointer to
+ *         a registry connection like app://known/direct/localhost.xml or
+ *         app://known/cognito/first_test.xml
  * 
  *         @auth: a java property file containing a username and password
- *         @index: the index to be used by harvest whose default is registry
- *         @trust_self_signed: all self signed certificates for https
- * 
- *         cognito: the cognito client ID, IDP and gateway for AWS based instances of opensearch
- *         server_url: the opensearch URL when not using AWS services
- *       
+ *        
  * 
  * <p>Java class for registry_type complex type.
  * 
@@ -36,16 +33,11 @@ import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
  * 
  * <pre>{@code
  * <complexType name="registry_type">
- *   <complexContent>
- *     <restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       <choice>
- *         <element name="cognitoCliendId" type="{}cognito_type"/>
- *         <element name="server_url" type="{}direct_type"/>
- *       </choice>
+ *   <simpleContent>
+ *     <extension base="<http://www.w3.org/2001/XMLSchema>normalizedString">
  *       <attribute name="auth" use="required" type="{http://www.w3.org/2001/XMLSchema}normalizedString" />
- *       <attribute name="index" type="{http://www.w3.org/2001/XMLSchema}normalizedString" default="registry" />
- *     </restriction>
- *   </complexContent>
+ *     </extension>
+ *   </simpleContent>
  * </complexType>
  * }</pre>
  * 
@@ -53,69 +45,41 @@ import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "registry_type", propOrder = {
-    "cognitoCliendId",
-    "serverUrl"
+    "value"
 })
 public class RegistryType {
 
-    protected CognitoType cognitoCliendId;
-    @XmlElement(name = "server_url")
-    protected DirectType serverUrl;
+    @XmlValue
+    @XmlJavaTypeAdapter(NormalizedStringAdapter.class)
+    @XmlSchemaType(name = "normalizedString")
+    protected String value;
     @XmlAttribute(name = "auth", required = true)
     @XmlJavaTypeAdapter(NormalizedStringAdapter.class)
     @XmlSchemaType(name = "normalizedString")
     protected String auth;
-    @XmlAttribute(name = "index")
-    @XmlJavaTypeAdapter(NormalizedStringAdapter.class)
-    @XmlSchemaType(name = "normalizedString")
-    protected String index;
 
     /**
-     * Gets the value of the cognitoCliendId property.
+     * Gets the value of the value property.
      * 
      * @return
      *     possible object is
-     *     {@link CognitoType }
+     *     {@link String }
      *     
      */
-    public CognitoType getCognitoCliendId() {
-        return cognitoCliendId;
+    public String getValue() {
+        return value;
     }
 
     /**
-     * Sets the value of the cognitoCliendId property.
+     * Sets the value of the value property.
      * 
      * @param value
      *     allowed object is
-     *     {@link CognitoType }
+     *     {@link String }
      *     
      */
-    public void setCognitoCliendId(CognitoType value) {
-        this.cognitoCliendId = value;
-    }
-
-    /**
-     * Gets the value of the serverUrl property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link DirectType }
-     *     
-     */
-    public DirectType getServerUrl() {
-        return serverUrl;
-    }
-
-    /**
-     * Sets the value of the serverUrl property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link DirectType }
-     *     
-     */
-    public void setServerUrl(DirectType value) {
-        this.serverUrl = value;
+    public void setValue(String value) {
+        this.value = value;
     }
 
     /**
@@ -140,34 +104,6 @@ public class RegistryType {
      */
     public void setAuth(String value) {
         this.auth = value;
-    }
-
-    /**
-     * Gets the value of the index property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getIndex() {
-        if (index == null) {
-            return "registry";
-        } else {
-            return index;
-        }
-    }
-
-    /**
-     * Sets the value of the index property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setIndex(String value) {
-        this.index = value;
     }
 
 }
